# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "manhtb1982@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Lấy danh sách products",
                "description": "Lấy danh sách products",
                "operationId": "indexProduct",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Phân trang",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "1"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tên sản phẩm",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Tạo product",
                "description": "Tạo product",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Product name",
                                        "type": "string",
                                        "example": "Mạnh"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "string",
                                        "example": "400"
                                    },
                                    "category_id": {
                                        "description": "Product category",
                                        "type": "string",
                                        "example": "4"
                                    },
                                    "color_id": {
                                        "description": "Product color",
                                        "type": "string",
                                        "enum": [
                                            "1",
                                            "2",
                                            "3",
                                            "4",
                                            "5"
                                        ]
                                    },
                                    "size_id": {
                                        "description": "Product size",
                                        "type": "string",
                                        "enum": [
                                            "1",
                                            "2",
                                            "3",
                                            "4",
                                            "5"
                                        ]
                                    },
                                    "img": {
                                        "description": "Product img",
                                        "type": "file"
                                    },
                                    "img_sub[]": {
                                        "description": "Product img_sub",
                                        "type": "file"
                                    },
                                    "description": {
                                        "description": "Product description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": "success"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Lấy chi tiết Product",
                "description": "Lấy chi tiết Product",
                "operationId": "showProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id find",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": "success",
                                        "data": {
                                            "id": 10,
                                            "name": "Áo Unisex 1",
                                            "price": 300,
                                            "color_id": "1,2,3,4",
                                            "size_id": "1,3,4",
                                            "category_id": 5,
                                            "img": "upload/img/img_251666578007.jpg",
                                            "description": "<p>Lorem ipsum</p>",
                                            "created_at": "2022-10-24T02:20:07.000000Z",
                                            "updated_at": "2022-10-24T02:20:07.000000Z"
                                        },
                                        "imgs": [
                                            {
                                                "id": 31,
                                                "path": "upload/img/10/img_1079410.jpg",
                                                "product_id": 10,
                                                "created_at": "2022-10-24T02:20:07.000000Z",
                                                "updated_at": "2022-10-24T02:20:07.000000Z"
                                            },
                                            {
                                                "id": 32,
                                                "path": "upload/img/10/img_395510.jpg",
                                                "product_id": 10,
                                                "created_at": "2022-10-24T02:20:07.000000Z",
                                                "updated_at": "2022-10-24T02:20:07.000000Z"
                                            },
                                            {
                                                "id": 33,
                                                "path": "upload/img/10/img_3641910.jpg",
                                                "product_id": 10,
                                                "created_at": "2022-10-24T02:20:07.000000Z",
                                                "updated_at": "2022-10-24T02:20:07.000000Z"
                                            },
                                            {
                                                "id": 34,
                                                "path": "upload/img/10/img_8871510.jpg",
                                                "product_id": 10,
                                                "created_at": "2022-10-24T02:20:08.000000Z",
                                                "updated_at": "2022-10-24T02:20:08.000000Z"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Xóa Product",
                "description": "Xóa Product",
                "operationId": "destroyProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id find",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "100"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": "success"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/getNumProducts": {
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Lấy số lượng sản phẩm theo danh mục",
                "description": "Lấy số lượng sản phẩm theo danh mục",
                "operationId": "getNumProducts",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "Id danh mục sản phẩm",
                                        "type": "string",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/products/search": {
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Tìm kiếm product",
                "description": "Tìm kiếm product",
                "operationId": "search",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "color": {
                                        "description": "Màu sản phẩm",
                                        "type": "string",
                                        "example": "0"
                                    },
                                    "size": {
                                        "description": "Size sản phẩm",
                                        "type": "string",
                                        "example": "0"
                                    },
                                    "category": {
                                        "description": "Danh mục sản phẩm",
                                        "type": "string",
                                        "example": "0"
                                    },
                                    "search_price": {
                                        "description": "Tìm kiếm theo giá",
                                        "type": "string",
                                        "example": "0"
                                    },
                                    "price_min": {
                                        "description": "Giá nhỏ nhất",
                                        "type": "string",
                                        "example": "100"
                                    },
                                    "price_max": {
                                        "description": "Giá lớn nhất",
                                        "type": "string",
                                        "example": "900"
                                    },
                                    "list_sort_post": {
                                        "description": "Sắp xếp giá",
                                        "type": "string",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}